Update README.md
Three mini games that help tie a premise together.

You, the player, is trying to save a world that is being ravished by its kind who is also the ruler of your world. The king collects life source from the other world to your world and leaves the inhabitants of that world to suffer.

You objective is to balance up the resources in both worlds, dethrone the King and vanish him in a final battle.

You accomplish this through three mini game. In the first mini game, you collect a life source. In the second, you transfer the life source through a portal and in the third mini game, you break a barricade through series of successive clicks and the finally, you dethrone the king and vanish him.