Demo Version
So we were able to do a demo of our game to the class. They liked it.
The presentation involved talking about our initial idea, the
prototyping process, the digital prototype playtest and the beta version
playtest.

We still have a lot of work to do with bringing together the platformer
and the mini games that I have already brought together.

I used a bunch of booleans and conditionals to start a game and end
others.

Our approach to this game is minimalistic. We use just darker and
related colors which I can easily change by changing the values in the
array of the color palette.